+++ pext/qml/AboutDialog.qml
@@ -80,6 +80,8 @@ Dialog {
                         "Emily Lau &lt;<a href='mailto:fuchslein@hackerchick.me'>fuchslein@hackerchick.me</a>&gt;<br><br>" +
                         "<b>" + qsTr("Hungarian") + "</b><br>" +
                         "Szöllősi Attila &lt;<a href='mailto:ata@openmailbox.org'>ata@openmailbox.org</a>&gt;<br><br>" +
+                        "<b>" + qsTr("Norwegian Bokmål") + "</b><br>" +
+                        "Allan Nordhøy &lt;<a href='mailto:epost@anotheragency.no'>epost@anotheragency.no</a>&gt;<br><br>" +
                         "<b>" + qsTr("Dutch") + "</b><br>" +
                         "Sylvia van Os &lt;<a href='mailto:sylvia@hackerchick.me'>sylvia@hackerchick.me</a>&gt;<br>" +
                         "Heimen Stoffels &lt;<a href='mailto:vistausss@outlook.com'>vistausss@outlook.com</a>&gt;<br><br>" +